Update documentation
I noted the sections still to come and the part I'm currently working 
on. Removed install instructions for operating systems I don't know if I 
will support or not.

Changing development methods and system to be less opinionated and more 
easily allow other developers to take part in developing the software.